removed assignment cycles actions and reducers
