--- src/reducers/assignmentCycle/index.js
@@ -1,13 +1,0 @@
-import {
-  assignmentCycle,
-  assignmentCycleFetchDataErrored,
-  assignmentCycleFetchDataLoading,
-  assignmentCycleSelections,
-} from './assignmentCycle';
-
-export default {
-  assignmentCycle,
-  assignmentCycleFetchDataErrored,
-  assignmentCycleFetchDataLoading,
-  assignmentCycleSelections,
-};
