+++ src/reducers/index.js
@@ -75,7 +75,6 @@ import PostPanelProcessing from './postPanelProcessing';
 import managePostAccess from './managePostAccess';
 import searchPostAccess from './searchPostAccess';
 import cycleJobCategories from './cycleJobCategories';
-import assignmentCycle from './assignmentCycle';
 import bureauException from './bureauException';
 import entryLevel from './entryLevel';
 import jobCategories from './jobCategories';
@@ -153,7 +152,6 @@ export default (history) => combineReducers({
   ...managePostAccess,
   ...searchPostAccess,
   ...cycleJobCategories,
-  ...assignmentCycle,
   ...bureauException,
   ...entryLevel,
   ...jobCategories,
